Fix size of arguments

+++ src/openrct2-ui/windows/ClearScenery.cpp
@@ -179,8 +179,8 @@ static void window_clear_scenery_textinput(rct_window* w, rct_widgetindex widget
 static void window_clear_scenery_inputsize(rct_window* w)
 {
     Formatter ft;
-    ft.Add<int32_t>(MINIMUM_TOOL_SIZE);
-    ft.Add<int32_t>(MAXIMUM_TOOL_SIZE);
+    ft.Add<int16_t>(MINIMUM_TOOL_SIZE);
+    ft.Add<int16_t>(MAXIMUM_TOOL_SIZE);
     window_text_input_open(w, WIDX_PREVIEW, STR_SELECTION_SIZE, STR_ENTER_SELECTION_SIZE, ft, STR_NONE, STR_NONE, 3);
 }
